Fix type definition
TypeScript is giving errors when using bitbar.separator, as there's some
error when refering to `unique symbol`s as properties of a `const`,
instead of exported variables of a `namespace`.

The rest of the changes are made to accomodate the typechecker merging a
function definition with a namespace which exports values (not only
types, as before).